Noncombining.

+++ ext/stats/kahan.go
@@ -5,15 +5,15 @@ package stats
 type kahan struct{ hi, lo float64 }
 
 func (k *kahan) add(x float64) {
-	y := k.lo + x
-	t := k.hi + y
-	k.lo = y - (t - k.hi)
+	y := float64(k.lo + x)
+	t := float64(k.hi + y)
+	k.lo = y - float64(t-k.hi)
 	k.hi = t
 }
 
 func (k *kahan) sub(x float64) {
-	y := k.lo - x
-	t := k.hi + y
-	k.lo = y - (t - k.hi)
+	y := float64(k.lo - x)
+	t := float64(k.hi + y)
+	k.lo = y - float64(t-k.hi)
 	k.hi = t
 }
